Added a lot of a creating function
Probably will have to scrap this function later - bc as far as I can tell it doesn't create any opportunities to work on the nonogram after it has been finished. This is a buggy update - the last section of the last function especially does not work, and the first part may have to be edited. This will be the last update for a while.